Upload developer guide and edit user guide
A developer guide is uploaded to explain the design of the script for developers. It also details which file or function to go in order to improve the script without understanding the whole design. 

The user guide is updated to mention a sample webserver and the format of web server log that users should follow.